+++ test/array.rb
@@ -6,11 +6,13 @@
 
 [1, 2, 3]
 
-# rubocop:disable Style/WordArray
-# We need to fix this, but it can't be done until we get the args_new and
-# args_add refactor done.
 ['a', 'b', 'c']
-# rubocop:enable Style/WordArray
+
+['a', 'b c', 'd']
+
+[:a, :b, :c]
+
+[:a, :"b c", :d]
 
 %w[a b c]